Story: abstracted instance methods into queries, mutations, shapers
