+++ models/story/story-instance-methods/story-instance-mutations.js
@@ -1,0 +1,11 @@
+async function publish() {
+  if (this.published) return null;
+
+  this.publishedAt = Date.now();
+  this.published = true;
+  return this.save();
+}
+
+module.exports = {
+  publish,
+};
